+++ src/main/java/domain/card/CardShuffler.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.ArrayList;
 import java.util.Collections;
