+++ src/main/java/domain/card/Pattern.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum Pattern {
