+++ src/main/java/domain/game/Players.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.game;
 
+import domain.card.CardDeck;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
