+++ src/main/java/view/OutputView.java
@@ -1,12 +1,12 @@
 package view;
 
-import domain.Card;
-import domain.CardDeck;
-import domain.CardNumber;
-import domain.Dealer;
-import domain.GameResult;
-import domain.Player;
-import domain.Players;
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.CardNumber;
+import domain.game.Dealer;
+import domain.game.GameResult;
+import domain.game.Player;
+import domain.game.Players;
 import java.util.List;
 
 public class OutputView {
