+++ src/test/java/domain/card/CardShufflerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.List;
 import org.assertj.core.api.Assertions;
