+++ src/test/java/domain/game/DealerTest.java
@@ -1,10 +1,82 @@
-package domain;
+package domain.game;
 
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.CardNumber;
+import domain.card.Pattern;
+import domain.card.TestShuffler;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class GameResultTest {
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class DealerTest {
+
+    @Test
+    void 딜러는_게임_시작_시_두_장의_카드를_받는다() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer();
+
+        //when
+        dealer.drawCardWhenStart(cardDeck);
+
+        //then
+        assertThat(dealer.getHand().getCards()).hasSize(2);
+    }
+
+    @Test
+    void 딜러는_한_장의_카드를_받는다() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer();
+
+        //when
+        dealer.drawCard(cardDeck);
+
+        //then
+        assertThat(dealer.getHand().getCards()).hasSize(1);
+    }
+
+    @Test
+    void 딜러가_보유한_카드의_합계를_구한다() {
+        //given
+        Dealer dealer = new Dealer();
+        Hand hand = dealer.getHand();
+        List<Card> cards = hand.getCards();
+        cards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        cards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
+
+        //when & then
+        assertThat(dealer.calculateTotalCardNumber()).isEqualTo(20);
+    }
+
+    @Test
+    void 딜러가_보유한_카드의_합계가_21을_넘었는지_판정한다() {
+        //given
+        Dealer dealer = new Dealer();
+        Hand hand = dealer.getHand();
+        List<Card> cards = hand.getCards();
+        cards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        cards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        //when & then
+        assertThat(dealer.isOverBurstBound()).isFalse();
+    }
+
+    @Test
+    void 딜러가_보유한_카드의_합계가_16을_넘었는지_판정한다() {
+        //given
+        Dealer dealer = new Dealer();
+        Hand hand = dealer.getHand();
+        List<Card> cards = hand.getCards();
+        cards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        cards.add(new Card(Pattern.CLOVER, CardNumber.SEVEN));
+
+        //when & then
+        assertThat(dealer.isOverDrawBound()).isTrue();
+    }
 
     @Test
     void 모든_플레이어들과_딜러_사이의_승패를_판정한다() {
@@ -22,10 +94,10 @@ public class GameResultTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.SIX));
 
         //when
-        GameResult gameResult = GameResult.of(dealer, player);
+        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
 
         //then
-        Assertions.assertThat(gameResult).isEqualTo(GameResult.LOSE);
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
     }
 
     @Test
@@ -46,10 +118,10 @@ public class GameResultTest {
         playerCards.add(new Card(Pattern.HEART, CardNumber.TEN));
 
         //when
-        GameResult gameResult = GameResult.of(dealer, player);
+        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
 
         //then
-        Assertions.assertThat(gameResult).isEqualTo(GameResult.LOSE);
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
     }
 
     @Test
@@ -69,10 +141,10 @@ public class GameResultTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
 
         //when
-        GameResult gameResult = GameResult.of(dealer, player);
+        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
 
         //then
-        Assertions.assertThat(gameResult).isEqualTo(GameResult.WIN);
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.WIN));
     }
 
     @Test
@@ -91,10 +163,10 @@ public class GameResultTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
 
         //when
-        GameResult gameResult = GameResult.of(dealer, player);
+        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
 
         //then
-        Assertions.assertThat(gameResult).isEqualTo(GameResult.DRAW);
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.DRAW));
     }
 
     @Test
@@ -114,10 +186,10 @@ public class GameResultTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
 
         //when
-        GameResult gameResult = GameResult.of(dealer, player);
+        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
 
         //then
-        Assertions.assertThat(gameResult).isEqualTo(GameResult.LOSE);
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
     }
 
     @Test
@@ -137,41 +209,9 @@ public class GameResultTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
 
         //when
-        GameResult gameResult = GameResult.of(dealer, player);
+        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
 
         //then
-        Assertions.assertThat(gameResult).isEqualTo(GameResult.WIN);
-    }
-
-    @Test
-    void 딜러의_승리_횟수를_계산한다() {
-        //when
-        int winCount = GameResult.WIN.countGameResult(
-                List.of(GameResult.WIN, GameResult.WIN, GameResult.LOSE, GameResult.DRAW));
-
-        //then
-        Assertions.assertThat(winCount).isEqualTo(1);
-
-    }
-
-    @Test
-    void 딜러의_패배_횟수를_계산한다() {
-        //when
-        int winCount = GameResult.LOSE.countGameResult(
-                List.of(GameResult.WIN, GameResult.WIN, GameResult.LOSE, GameResult.DRAW));
-
-        //then
-        Assertions.assertThat(winCount).isEqualTo(2);
-
-    }
-
-    @Test
-    void 딜러의_무승부_횟수를_계산한다() {
-        //when
-        int winCount = GameResult.DRAW.countGameResult(
-                List.of(GameResult.WIN, GameResult.WIN, GameResult.LOSE, GameResult.DRAW));
-
-        //then
-        Assertions.assertThat(winCount).isEqualTo(1);
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.WIN));
     }
 }
