+++ src/test/java/domain/game/GameResultTest.java
@@ -1,77 +1,13 @@
-package domain;
+package domain.game;
 
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.Pattern;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class DealerTest {
-
-    @Test
-    void 딜러는_게임_시작_시_두_장의_카드를_받는다() {
-        //given
-        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
-        Dealer dealer = new Dealer();
-
-        //when
-        dealer.drawCardWhenStart(cardDeck);
-
-        //then
-        assertThat(dealer.getHand().getCards()).hasSize(2);
-    }
-
-    @Test
-    void 딜러는_한_장의_카드를_받는다() {
-        //given
-        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
-        Dealer dealer = new Dealer();
-
-        //when
-        dealer.drawCard(cardDeck);
-
-        //then
-        assertThat(dealer.getHand().getCards()).hasSize(1);
-    }
-
-    @Test
-    void 딜러가_보유한_카드의_합계를_구한다() {
-        //given
-        Dealer dealer = new Dealer();
-        Hand hand = dealer.getHand();
-        List<Card> cards = hand.getCards();
-        cards.add(new Card(Pattern.SPADE, CardNumber.TEN));
-        cards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
-
-        //when & then
-        assertThat(dealer.calculateTotalCardNumber()).isEqualTo(20);
-    }
-
-    @Test
-    void 딜러가_보유한_카드의_합계가_21을_넘었는지_판정한다() {
-        //given
-        Dealer dealer = new Dealer();
-        Hand hand = dealer.getHand();
-        List<Card> cards = hand.getCards();
-        cards.add(new Card(Pattern.SPADE, CardNumber.TEN));
-        cards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
-
-        //when & then
-        assertThat(dealer.isOverBurstBound()).isFalse();
-    }
-
-    @Test
-    void 딜러가_보유한_카드의_합계가_16을_넘었는지_판정한다() {
-        //given
-        Dealer dealer = new Dealer();
-        Hand hand = dealer.getHand();
-        List<Card> cards = hand.getCards();
-        cards.add(new Card(Pattern.SPADE, CardNumber.TEN));
-        cards.add(new Card(Pattern.CLOVER, CardNumber.SEVEN));
-
-        //when & then
-        assertThat(dealer.isOverDrawBound()).isTrue();
-    }
+public class GameResultTest {
 
     @Test
     void 모든_플레이어들과_딜러_사이의_승패를_판정한다() {
@@ -89,10 +25,10 @@ public class DealerTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.SIX));
 
         //when
-        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
+        GameResult gameResult = GameResult.of(dealer, player);
 
         //then
-        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
+        Assertions.assertThat(gameResult).isEqualTo(GameResult.LOSE);
     }
 
     @Test
@@ -113,10 +49,10 @@ public class DealerTest {
         playerCards.add(new Card(Pattern.HEART, CardNumber.TEN));
 
         //when
-        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
+        GameResult gameResult = GameResult.of(dealer, player);
 
         //then
-        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
+        Assertions.assertThat(gameResult).isEqualTo(GameResult.LOSE);
     }
 
     @Test
@@ -136,10 +72,10 @@ public class DealerTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
 
         //when
-        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
+        GameResult gameResult = GameResult.of(dealer, player);
 
         //then
-        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.WIN));
+        Assertions.assertThat(gameResult).isEqualTo(GameResult.WIN);
     }
 
     @Test
@@ -158,10 +94,10 @@ public class DealerTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
 
         //when
-        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
+        GameResult gameResult = GameResult.of(dealer, player);
 
         //then
-        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.DRAW));
+        Assertions.assertThat(gameResult).isEqualTo(GameResult.DRAW);
     }
 
     @Test
@@ -181,10 +117,10 @@ public class DealerTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
 
         //when
-        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
+        GameResult gameResult = GameResult.of(dealer, player);
 
         //then
-        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
+        Assertions.assertThat(gameResult).isEqualTo(GameResult.LOSE);
     }
 
     @Test
@@ -204,9 +140,41 @@ public class DealerTest {
         playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
 
         //when
-        List<GameResult> gameResult = dealer.judgeGameResult(List.of(player));
+        GameResult gameResult = GameResult.of(dealer, player);
 
         //then
-        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.WIN));
+        Assertions.assertThat(gameResult).isEqualTo(GameResult.WIN);
+    }
+
+    @Test
+    void 딜러의_승리_횟수를_계산한다() {
+        //when
+        int winCount = GameResult.WIN.countGameResult(
+                List.of(GameResult.WIN, GameResult.WIN, GameResult.LOSE, GameResult.DRAW));
+
+        //then
+        Assertions.assertThat(winCount).isEqualTo(1);
+
+    }
+
+    @Test
+    void 딜러의_패배_횟수를_계산한다() {
+        //when
+        int winCount = GameResult.LOSE.countGameResult(
+                List.of(GameResult.WIN, GameResult.WIN, GameResult.LOSE, GameResult.DRAW));
+
+        //then
+        Assertions.assertThat(winCount).isEqualTo(2);
+
+    }
+
+    @Test
+    void 딜러의_무승부_횟수를_계산한다() {
+        //when
+        int winCount = GameResult.DRAW.countGameResult(
+                List.of(GameResult.WIN, GameResult.WIN, GameResult.LOSE, GameResult.DRAW));
+
+        //then
+        Assertions.assertThat(winCount).isEqualTo(1);
     }
 }
